Move discovery app to discovery_legacy to prepare for the new disco pane
Note that the only model, DiscoveryModule, already specified a custom
db_table so we don't need to modify that.
