+++ src/olympia/legacy_discovery/admin.py
