+++ src/olympia/legacy_discovery/helpers.py
