+++ src/olympia/legacy_discovery/templates/legacy_discovery/addons/base.html
@@ -1,4 +1,4 @@
-{% extends "discovery/base.html" %}
+{% extends "legacy_discovery/base.html" %}
 
 {% block bodyclass %}detail{% endblock %}
