+++ src/olympia/legacy_discovery/templates/legacy_discovery/addons/detail.html
@@ -1,4 +1,4 @@
-{% extends "discovery/addons/base.html" %}
+{% extends "legacy_discovery/addons/base.html" %}
 
 {% block main %}
 {% if addon.type == amo.ADDON_PERSONA %}
