+++ src/olympia/legacy_discovery/templates/legacy_discovery/addons/eula.html
@@ -1,4 +1,4 @@
-{% extends "discovery/addons/base.html" %}
+{% extends "legacy_discovery/addons/base.html" %}
 
 {% block title %}
   {# L10n: {0} is the name of the add-on #}
