+++ src/olympia/legacy_discovery/templates/legacy_discovery/module_admin.html
