+++ src/olympia/legacy_discovery/templates/legacy_discovery/modules/collection.html
