+++ src/olympia/legacy_discovery/templates/legacy_discovery/modules/featured.html
