+++ src/olympia/legacy_discovery/templates/legacy_discovery/more_addons.html
@@ -1,4 +1,4 @@
-{% from "discovery/macros.html" import addon_list with context %}
+{% from "legacy_discovery/macros.html" import addon_list with context %}
 
 {% if featured_addons %}
   {{ addon_list(featured_addons, limit=20) }}
