+++ src/olympia/legacy_discovery/templates/legacy_discovery/pane.html
@@ -1,5 +1,5 @@
-{% extends "discovery/base.html" %}
-{% from "discovery/macros.html" import addon_list, persona_list with context %}
+{% extends "legacy_discovery/base.html" %}
+{% from "legacy_discovery/macros.html" import addon_list, persona_list with context %}
 
 {% set app = request.APP.pretty %}
