+++ src/olympia/legacy_discovery/templates/legacy_discovery/pane_account.html
