+++ src/olympia/lib/settings_base.py
@@ -377,11 +377,11 @@ INSTALLED_APPS = (
     'olympia.browse',
     'olympia.compat',
     'olympia.devhub',
-    'olympia.discovery',
     'olympia.editors',
     'olympia.files',
     'olympia.internal_tools',
     'olympia.legacy_api',
+    'olympia.legacy_discovery',
     'olympia.lib.es',
     'olympia.pages',
     'olympia.perf',
